requirements.txt
pytorch
tifffile
napari-ndtiffs